% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.budgets_operations.R
\name{create_budget}
\alias{create_budget}
\title{Creates a budget and, if included, notifications and subscribers}
\usage{
create_budget(AccountId, Budget, NotificationsWithSubscribers = NULL)
}
\arguments{
\item{AccountId}{The \code{accountId} that is associated with the budget.}

\item{Budget}{The budget object that you want to create.}

\item{NotificationsWithSubscribers}{A notification that you want to associate with a budget. A budget can have up to five notifications, and each notification can have one SNS subscriber and up to ten email subscribers. If you include notifications and subscribers in your \code{CreateBudget} call, AWS creates the notifications and subscribers for you.}
}
\description{
Creates a budget and, if included, notifications and subscribers.
}
\examples{

}
