% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.budgets_operations.R
\name{create_subscriber}
\alias{create_subscriber}
\title{Creates a subscriber}
\usage{
create_subscriber(AccountId, BudgetName, Notification, Subscriber)
}
\arguments{
\item{AccountId}{The \code{accountId} associated with the budget that you want to create a subscriber for.}

\item{BudgetName}{The name of the budget that you want to subscribe to. Budget names must be unique within an account.}

\item{Notification}{The notification that you want to create a subscriber for.}

\item{Subscriber}{The subscriber that you want to associate with a budget notification.}
}
\description{
Creates a subscriber. You must create the associated budget and notification before you create the subscriber.
}
\examples{

}
