% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.budgets_operations.R
\name{delete_budget}
\alias{delete_budget}
\title{Deletes a budget}
\usage{
delete_budget(AccountId, BudgetName)
}
\arguments{
\item{AccountId}{The \code{accountId} that is associated with the budget that you want to delete.}

\item{BudgetName}{The name of the budget that you want to delete.}
}
\description{
Deletes a budget. You can delete your budget at any time.
}
\details{
\strong{Deleting a budget also deletes the notifications and subscribers associated with that budget.}
}
\examples{

}
