% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.budgets_operations.R
\name{delete_notification}
\alias{delete_notification}
\title{Deletes a notification}
\usage{
delete_notification(AccountId, BudgetName, Notification)
}
\arguments{
\item{AccountId}{The \code{accountId} that is associated with the budget whose notification you want to delete.}

\item{BudgetName}{The name of the budget whose notification you want to delete.}

\item{Notification}{The notification that you want to delete.}
}
\description{
Deletes a notification.
}
\details{
\strong{Deleting a notification also deletes the subscribers associated with the notification.}
}
\examples{

}
