% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.budgets_operations.R
\name{delete_subscriber}
\alias{delete_subscriber}
\title{Deletes a subscriber}
\usage{
delete_subscriber(AccountId, BudgetName, Notification, Subscriber)
}
\arguments{
\item{AccountId}{The \code{accountId} that is associated with the budget whose subscriber you want to delete.}

\item{BudgetName}{The name of the budget whose subscriber you want to delete.}

\item{Notification}{The notification whose subscriber you want to delete.}

\item{Subscriber}{The subscriber that you want to delete.}
}
\description{
Deletes a subscriber.
}
\details{
\strong{Deleting the last subscriber to a notification also deletes the notification.}
}
\examples{

}
