% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.budgets_operations.R
\name{describe_budget}
\alias{describe_budget}
\title{Describes a budget}
\usage{
describe_budget(AccountId, BudgetName)
}
\arguments{
\item{AccountId}{The \code{accountId} that is associated with the budget that you want a description of.}

\item{BudgetName}{The name of the budget that you want a description of.}
}
\description{
Describes a budget.
}
\examples{

}
