% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.budgets_operations.R
\name{describe_budget_performance_history}
\alias{describe_budget_performance_history}
\title{Describes the history for DAILY, MONTHLY, and QUARTERLY budgets}
\usage{
describe_budget_performance_history(AccountId, BudgetName,
  TimePeriod = NULL, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{AccountId}{[required]}

\item{BudgetName}{[required]}

\item{TimePeriod}{Retrieves how often the budget went into an \code{ALARM} state for the specified time period.}

\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Describes the history for \code{DAILY}, \code{MONTHLY}, and \code{QUARTERLY} budgets. Budget history isn't available for \code{ANNUAL} budgets.
}
\section{Accepted Parameters}{
\preformatted{describe_budget_performance_history(
  AccountId = "string",
  BudgetName = "string",
  TimePeriod = list(
    Start = as.POSIXct("2015-01-01"),
    End = as.POSIXct("2015-01-01")
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

