% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.budgets_operations.R
\name{describe_budgets}
\alias{describe_budgets}
\title{Lists the budgets associated with an account}
\usage{
describe_budgets(AccountId, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{AccountId}{The \code{accountId} that is associated with the budgets that you want descriptions of.}

\item{MaxResults}{Optional integer. Specifies the maximum number of results to return in response.}

\item{NextToken}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Lists the budgets associated with an account.
}
\examples{

}
