% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.budgets_operations.R
\name{update_budget}
\alias{update_budget}
\title{Updates a budget}
\usage{
update_budget(AccountId, NewBudget)
}
\arguments{
\item{AccountId}{The \code{accountId} that is associated with the budget that you want to update.}

\item{NewBudget}{The budget that you want to update your budget to.}
}
\description{
Updates a budget. You can change every part of a budget except for the \code{budgetName} and the \code{calculatedSpend}. When a budget is modified, the \code{calculatedSpend} drops to zero until AWS has new usage data to use for forecasting.
}
\examples{

}
