% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.budgets_operations.R
\name{update_subscriber}
\alias{update_subscriber}
\title{Updates a subscriber}
\usage{
update_subscriber(AccountId, BudgetName, Notification, OldSubscriber,
  NewSubscriber)
}
\arguments{
\item{AccountId}{The \code{accountId} that is associated with the budget whose subscriber you want to update.}

\item{BudgetName}{The name of the budget whose subscriber you want to update.}

\item{Notification}{The notification whose subscriber you want to update.}

\item{OldSubscriber}{The previous subscriber associated with a budget notification.}

\item{NewSubscriber}{The updated subscriber associated with a budget notification.}
}
\description{
Updates a subscriber.
}
\examples{

}
