% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.budgets_operations.R
\name{create_notification}
\alias{create_notification}
\title{Creates a notification}
\usage{
create_notification(AccountId, BudgetName, Notification, Subscribers)
}
\arguments{
\item{AccountId}{The \code{accountId} that is associated with the budget that you want to create a notification for.}

\item{BudgetName}{The name of the budget that you want AWS to notified you about. Budget names must be unique within an account.}

\item{Notification}{The notification that you want to create.}

\item{Subscribers}{A list of subscribers that you want to associate with the notification. Each notification can have one SNS subscriber and up to ten email subscribers.}
}
\description{
Creates a notification. You must create the budget before you create the associated notification.
}
\examples{

}
