% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.budgets_operations.R
\name{delete_budget}
\alias{delete_budget}
\title{Deletes a budget}
\usage{
delete_budget(AccountId, BudgetName)
}
\arguments{
\item{AccountId}{[required] The \code{accountId} that is associated with the budget that you want to delete.}

\item{BudgetName}{[required] The name of the budget that you want to delete.}
}
\description{
Deletes a budget. You can delete your budget at any time.
}
\details{
Deleting a budget also deletes the notifications and subscribers that are associated with that budget.
}
\section{Accepted Parameters}{
\preformatted{delete_budget(
  AccountId = "string",
  BudgetName = "string"
)
}
}

