% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.budgets_operations.R
\name{describe_subscribers_for_notification}
\alias{describe_subscribers_for_notification}
\title{Lists the subscribers associated with a notification}
\usage{
describe_subscribers_for_notification(AccountId, BudgetName, Notification,
  MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{AccountId}{The \code{accountId} that is associated with the budget whose subscribers you want descriptions of.}

\item{BudgetName}{The name of the budget whose subscribers you want descriptions of.}

\item{Notification}{The notification whose subscribers you want to list.}

\item{MaxResults}{Optional integer. Specifies the maximum number of results to return in response.}

\item{NextToken}{The pagination token that indicates the next set of results to retrieve.}
}
\description{
Lists the subscribers associated with a notification.
}
\examples{

}
