% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.budgets_package.R
\docType{package}
\name{paws.budgets-package}
\alias{paws.budgets}
\alias{paws.budgets-package}
\title{paws.budgets: AWS Budgets}
\description{
- Cost budgets allow you to say how much you
want to spend on a service.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
