% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.budgets_operations.R
\name{update_notification}
\alias{update_notification}
\title{Updates a notification}
\usage{
update_notification(AccountId, BudgetName, OldNotification, NewNotification)
}
\arguments{
\item{AccountId}{The \code{accountId} that is associated with the budget whose notification you want to update.}

\item{BudgetName}{The name of the budget whose notification you want to update.}

\item{OldNotification}{The previous notification associated with a budget.}

\item{NewNotification}{The updated notification to be associated with a budget.}
}
\description{
Updates a notification.
}
\examples{

}
